исправил часть замечаний
